Adds logout functionality to the application
Implements a "Logout" option in the application menu.
This allows users to explicitly log out of their accounts.

The logout action sends a "logout" event to the renderer process,
which can then handle the necessary cleanup and redirection.

Also updates the application version to 0.2.1 and adds author and description information to the package.json file.